Print container logs if port wait timeout is exceeded while initializing metastore backend
There are many cases where reinit_metastore script fails while waiting for a database port to become ready.
When the port timeout is reached the following message is displayed:
```
[2023-03-30T19:38:16.613Z] waiting for oracle to be available...
[2023-03-30T19:43:23.540Z] timeout reached before the port went into state "inuse"
script returned exit code 1
```
but we don't have any clue on what is happening inside the container during that time.

With this change the container logs will be displayed just before exiting which can hopefully give more insights on what's going on inside the container.